Search for baselearner in this package
Search for `baselearner` in this package, not in the caller's namespace.

When `baselearner` is given as a string, it's limited (by `match.arg`) to the pre-defined choices `c("bbs", "bols", "btree", "bss", "bns")`.  It doesn't make sense to require the caller to import that symbol when we know it's defined in the `mboost` package anyway.